+++ lib/src/platform_extensions.dart
@@ -1,0 +1,11 @@
+import 'dart:io';
+
+extension TPlatform on Platform {
+  static bool get isDesktop {
+    return Platform.isLinux || Platform.isMacOS || Platform.isWindows;
+  }
+
+  static bool get isMobile {
+    return Platform.isAndroid || Platform.isIOS;
+  }
+}