+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# 0.2.0
+# 0.3.0
+
+- Added `DateTime.calculateAge`
+- Added `DateTime.calculateDetailedAge`
+- Added `DateTime.toDetailedAgeLabel`
+- Added `DateTime.formatTime`
+- Added `Duration.toRemainingLabel`
+- Added `Duration.testExecutionTime`
+- Added `DurationEtaExtensions.testExecutionTime`
+
+## 0.2.0
 
 - Added `MapCoreExtensions.putMapList`
 - Added `MapCoreExtensions.getMapList`